Largely rewrite SourceTree.py to make easier to understand
There's a few minor behaviour changes: paths aren't compared with
the various patterns case-insensitively now (this effects nothing
in csswg-test), and the exact syntax of reftest filenames has
changed (this fixes mediaqueries-3/ref.html so that it is picked
up as a reference).

This keeps the existing SourceTree class API for the moment. We
should probably eventually drop it, as it really doesn't make
sense for it to be a class.